pyproject.toml
[tool.poetry]
name = "supplychain"
version = "0.1.0"
description = ""
authors = ["Yerzhan Amirkhanov <yerzhanamirkhanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
